Simplify rest of the interface
