+++ sdk/deepbook/src/client.ts
@@ -17,13 +17,22 @@ import {
 } from '@mysten/sui.js/utils';
 import { SuiClient, getFullnodeUrl } from '@mysten/sui.js/client';
 import { TransactionBlock } from '@mysten/sui.js/transactions';
-import { MODULE_CLOB, PACKAGE_ID, NORMALIZED_SUI_COIN_TYPE } from './utils';
-import { PaginatedPoolSummary, PoolSummary, UserPosition } from './types/pool';
+import {
+	MODULE_CLOB,
+	PACKAGE_ID,
+	NORMALIZED_SUI_COIN_TYPE,
+	FLOAT_SCALING,
+	CREATION_FEE,
+	MODULE_CUSTODIAN,
+	ORDER_DEFAULT_EXPIRATION_IN_MS,
+} from './utils';
+import { LimitOrderType, PaginatedPoolSummary, PoolSummary, UserPosition } from './types/pool';
 import { Coin } from '@mysten/sui.js';
 
 const DUMMY_ADDRESS = normalizeSuiAddress('0x0');
 
 export class DeepBookClient {
+	#poolTypeArgsCache: Map<string, string[]> = new Map();
 	/**
 	 *
 	 * @param suiClient connection to fullnode
@@ -34,6 +43,7 @@ export class DeepBookClient {
 		public suiClient: SuiClient = new SuiClient({ url: getFullnodeUrl('testnet') }),
 		public accountCap: string | undefined = undefined,
 		public currentAddress: string = DUMMY_ADDRESS,
+		private clientOrderId: number = 0,
 	) {}
 
 	/**
@@ -41,6 +51,98 @@ export class DeepBookClient {
 	 */
 	async setAccountCap(cap: string) {
 		this.accountCap = cap;
+	}
+
+	/**
+	 * @description: Create pool for trading pair
+	 * @param baseAssetType Full coin type of the base asset, eg: "0x3d0d0ce17dcd3b40c2d839d96ce66871ffb40e1154a8dd99af72292b3d10d7fc::wbtc::WBTC"
+	 * @param quoteAssetType Full coin type of quote asset, eg: "0x3d0d0ce17dcd3b40c2d839d96ce66871ffb40e1154a8dd99af72292b3d10d7fc::usdt::USDT"
+	 * @param tickSize Minimal Price Change Accuracy of this pool, eg: 10000000
+	 * @param lotSize Minimal Lot Change Accuracy of this pool, eg: 10000
+	 */
+	createPool(
+		baseAssetType: string,
+		quoteAssetType: string,
+		tickSize: number,
+		lotSize: number,
+	): TransactionBlock {
+		const txb = new TransactionBlock();
+		// create a pool with CREATION_FEE
+		const [coin] = txb.splitCoins(txb.gas, [txb.pure(CREATION_FEE)]);
+		txb.moveCall({
+			typeArguments: [baseAssetType, quoteAssetType],
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::create_pool`,
+			arguments: [txb.pure(tickSize), txb.pure(lotSize), coin],
+		});
+		return txb;
+	}
+
+	/**
+	 * @description: Create pool for trading pair
+	 * @param baseAssetType Full coin type of the base asset, eg: "0x3d0d0ce17dcd3b40c2d839d96ce66871ffb40e1154a8dd99af72292b3d10d7fc::wbtc::WBTC"
+	 * @param quoteAssetType Full coin type of quote asset, eg: "0x3d0d0ce17dcd3b40c2d839d96ce66871ffb40e1154a8dd99af72292b3d10d7fc::usdt::USDT"
+	 * @param tickSize Minimal Price Change Accuracy of this pool, eg: 10000000
+	 * @param lotSize Minimal Lot Change Accuracy of this pool, eg: 10000
+	 * @param takerFeeRate Customized taker fee rate, 10^9 scaling, Taker_fee_rate of 0.25% should be 2_500_000 for example
+	 * @param makerRebateRate Customized maker rebate rate, 10^9 scaling,  should be less than or equal to the taker_rebate_rate
+	 */
+	createCustomizedPool(
+		baseAssetType: string,
+		quoteAssetType: string,
+		tickSize: number,
+		lotSize: number,
+		takerFeeRate: number,
+		makerRebateRate: number,
+	): TransactionBlock {
+		const txb = new TransactionBlock();
+		// create a pool with CREATION_FEE
+		const [coin] = txb.splitCoins(txb.gas, [txb.pure(CREATION_FEE)]);
+		txb.moveCall({
+			typeArguments: [baseAssetType, quoteAssetType],
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::create_customized_pool`,
+			arguments: [
+				txb.pure(tickSize),
+				txb.pure(lotSize),
+				txb.pure(takerFeeRate),
+				txb.pure(makerRebateRate),
+				coin,
+			],
+		});
+		return txb;
+	}
+
+	/**
+	 * @description: Create and Transfer custodian account to user
+	 * @param currentAddress: current user address, eg: "0xbddc9d4961b46a130c2e1f38585bbc6fa8077ce54bcb206b26874ac08d607966"
+	 */
+	createAccount(currentAddress: string = this.currentAddress): TransactionBlock {
+		const txb = new TransactionBlock();
+		let [cap] = txb.moveCall({
+			typeArguments: [],
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::create_account`,
+			arguments: [],
+		});
+		txb.transferObjects([cap], txb.pure(this.#checkAddress(currentAddress)));
+		return txb;
+	}
+
+	/**
+	 * @description: Create and Transfer custodian account to user
+	 * @param currentAddress: current user address, eg: "0xbddc9d4961b46a130c2e1f38585bbc6fa8077ce54bcb206b26874ac08d607966"
+	 * @param accountCap: Object id of Account Capacity under user address, created after invoking createAccount, eg: "0x6f699fef193723277559c8f499ca3706121a65ac96d273151b8e52deb29135d3"
+	 */
+	createChildAccountCap(
+		currentAddress: string = this.currentAddress,
+		accountCap: string | undefined = this.accountCap,
+	): TransactionBlock {
+		const txb = new TransactionBlock();
+		let [childCap] = txb.moveCall({
+			typeArguments: [],
+			target: `${PACKAGE_ID}::${MODULE_CUSTODIAN}::create_child_account_cap`,
+			arguments: [txb.object(this.#checkAccountCap(accountCap))],
+		});
+		txb.transferObjects([childCap], txb.pure(this.#checkAddress(currentAddress)));
+		return txb;
 	}
 
 	/**
@@ -53,11 +155,11 @@ export class DeepBookClient {
 	async deposit(
 		poolId: string,
 		coinId: string | undefined = undefined,
-		amount: bigint | number | undefined = undefined,
+		quantity: bigint | number | undefined = undefined,
 	): Promise<TransactionBlock> {
 		const txb = new TransactionBlock();
 
-		const { baseAsset, quoteAsset } = await this.getPoolInfo(poolId);
+		const [baseAsset, quoteAsset] = await this.getPoolTypeArgs(poolId);
 		const hasSui =
 			baseAsset === NORMALIZED_SUI_COIN_TYPE || quoteAsset === NORMALIZED_SUI_COIN_TYPE;
 
@@ -67,7 +169,7 @@ export class DeepBookClient {
 
 		const inputCoin = coinId ? txb.object(coinId) : txb.gas;
 
-		const [coin] = amount ? txb.splitCoins(inputCoin, [txb.pure(amount)]) : [inputCoin];
+		const [coin] = quantity ? txb.splitCoins(inputCoin, [txb.pure(quantity)]) : [inputCoin];
 
 		const coinType = coinId ? this.#getCoinType(coinId) : NORMALIZED_SUI_COIN_TYPE;
 		if (coinType !== baseAsset && coinType !== quoteAsset) {
@@ -84,7 +186,7 @@ export class DeepBookClient {
 	}
 
 	/**
-	 * @description construct transaction block for withdrawing asset into a pool.
+	 * @description construct transaction block for withdrawing asset from a pool.
 	 * @param poolId the pool id for the withdraw
 	 * @param amount the amount of coin to withdraw
 	 * @param assetType Base or Quote
@@ -94,22 +196,246 @@ export class DeepBookClient {
 	async withdraw(
 		poolId: string,
 		// TODO: implement withdraw all
-		amount: bigint | number,
+		quantity: bigint | number,
 		assetType: 'Base' | 'Quote',
 		recipientAddress: string = this.currentAddress,
 	): Promise<TransactionBlock> {
 		const txb = new TransactionBlock();
-
-		const { baseAsset, quoteAsset } = await this.getPoolInfo(poolId);
-
 		const functionName = assetType === 'Base' ? 'withdraw_base' : 'withdraw_quote';
-
 		const [withdraw] = txb.moveCall({
-			typeArguments: [baseAsset, quoteAsset],
+			typeArguments: await this.getPoolTypeArgs(poolId),
 			target: `${PACKAGE_ID}::${MODULE_CLOB}::${functionName}`,
-			arguments: [txb.object(poolId), txb.pure(amount), txb.object(this.#checkAccountCap())],
+			arguments: [txb.object(poolId), txb.pure(quantity), txb.object(this.#checkAccountCap())],
 		});
-		txb.transferObjects([withdraw], txb.pure(this.#checkCurrentAddress(recipientAddress)));
+		txb.transferObjects([withdraw], txb.pure(this.#checkAddress(recipientAddress)));
+		return txb;
+	}
+
+	/**
+	 * @description: place a limit order
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param price: price of the limit order, eg: 180000000
+	 * @param quantity: quantity of the limit order in BASE ASSET, eg: 100000000
+	 * @param isBid: true for buying base with quote, false for selling base for quote
+	 * @param expirationTimestamp: expiration timestamp of the limit order in ms, eg: 1620000000000. If omitted, the order will expire in 1 day
+	 * from the time this function is called(not the time the transaction is executed)
+	 * @param restriction restrictions on limit orders, explain in doc for more details, eg: 0
+	 * @param clientOrderId a client side defined order number for bookkeeping purpose, e.g., "1", "2", etc. If omitted, the sdk will
+	 * assign a increasing number starting from 0. But this number might be duplicated if you are using multiple sdk instances
+	 * @param selfMatchingPrevention: whether to use self matching prevention. If omitted, the default value is true
+	 */
+	async placeLimitOrder(
+		poolId: string,
+		price: number,
+		quantity: bigint | number,
+		isBid: boolean,
+		expirationTimestamp: number | undefined = Date.now() + ORDER_DEFAULT_EXPIRATION_IN_MS,
+		orderType: LimitOrderType = LimitOrderType.NO_RESTRICTION,
+		clientOrderId: string | undefined,
+		selfMatchingPrevention: boolean = true,
+	): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		const args = [
+			txb.object(poolId),
+			txb.pure(clientOrderId ?? this.#nextClientOrderId()),
+			txb.pure(Math.floor(price * FLOAT_SCALING)), // to avoid float number
+			txb.pure(quantity),
+			txb.pure(selfMatchingPrevention),
+			txb.pure(isBid),
+			txb.pure(expirationTimestamp),
+			txb.pure(orderType),
+			txb.object(SUI_CLOCK_OBJECT_ID),
+			txb.object(this.#checkAccountCap()),
+		];
+		txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::place_limit_order`,
+			arguments: args,
+		});
+		return txb;
+	}
+
+	/**
+	 * @description: place a market order
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param quantity Amount of quote asset to swap in base asset
+	 * @param isBid true if the order is bid, false if the order is ask
+	 * @param baseCoin the objectId of the base coin
+	 * @param quoteCoin the objectId of the quote coin
+	 * @param clientOrderId a client side defined order id for bookkeeping purpose. eg: "1" , "2", ... If omitted, the sdk will
+	 * assign a increasing number starting from 0. But this number might be duplicated if you are using multiple sdk instances
+	 * @param recipientAddress: address to return the unused amounts, eg: "0xbddc9d4961b46a130c2e1f38585bbc6fa8077ce54bcb206b26874ac08d607966"
+	 */
+	async placeMarketOrder(
+		poolId: string,
+		quantity: bigint | number,
+		isBid: boolean,
+		baseCoin: string,
+		quoteCoin: string,
+		clientOrderId: string | undefined = undefined,
+		recipientAddress: string = this.currentAddress,
+	): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		const [base_coin_ret, quote_coin_ret] = txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::place_market_order`,
+			arguments: [
+				txb.object(poolId),
+				txb.object(this.#checkAccountCap()),
+				txb.pure(clientOrderId ?? this.#nextClientOrderId()),
+				txb.pure(quantity),
+				txb.pure(isBid),
+				txb.object(baseCoin),
+				txb.object(quoteCoin),
+				txb.object(SUI_CLOCK_OBJECT_ID),
+			],
+		});
+		const recipient = this.#checkAddress(recipientAddress);
+		txb.transferObjects([base_coin_ret], txb.pure(recipient));
+		txb.transferObjects([quote_coin_ret], txb.pure(recipient));
+		return txb;
+	}
+
+	/**
+	 * @description: swap exact quote for base
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param tokenObjectIn: Object id of the token to swap: eg: "0x6e566fec4c388eeb78a7dab832c9f0212eb2ac7e8699500e203def5b41b9c70d"
+	 * @param amountIn: amount of token to buy or sell, eg: 10000000
+	 * @param currentAddress: current user address, eg: "0xbddc9d4961b46a130c2e1f38585bbc6fa8077ce54bcb206b26874ac08d607966"
+	 * @param clientOrderId a client side defined order id for bookkeeping purpose, eg: "1" , "2", ... If omitted, the sdk will
+	 * assign a increasing number starting from 0. But this number might be duplicated if you are using multiple sdk instances
+	 */
+	async swapExactQuoteForBase(
+		poolId: string,
+		tokenObjectIn: string,
+		amountIn: bigint | number,
+		currentAddress: string,
+		clientOrderId: string | undefined = undefined,
+	): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		// in this case, we assume that the tokenIn--tokenOut always exists.
+		const [base_coin_ret, quote_coin_ret, _amount] = txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::swap_exact_quote_for_base`,
+			arguments: [
+				txb.object(poolId),
+				txb.pure(clientOrderId ?? this.#nextClientOrderId()),
+				txb.object(this.#checkAccountCap()),
+				txb.object(String(amountIn)),
+				txb.object(SUI_CLOCK_OBJECT_ID),
+				txb.object(tokenObjectIn),
+			],
+		});
+		txb.transferObjects([base_coin_ret], txb.pure(currentAddress));
+		txb.transferObjects([quote_coin_ret], txb.pure(currentAddress));
+		return txb;
+	}
+
+	/**
+	 * @description swap exact base for quote
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param tokenObjectIn Object id of the token to swap: eg: "0x6e566fec4c388eeb78a7dab832c9f0212eb2ac7e8699500e203def5b41b9c70d"
+	 * @param amountIn amount of token to buy or sell, eg: 10000000
+	 * @param currentAddress current user address, eg: "0xbddc9d4961b46a130c2e1f38585bbc6fa8077ce54bcb206b26874ac08d607966"
+	 * @param clientOrderId a client side defined order number for bookkeeping purpose. eg: "1" , "2", ...
+	 */
+	async swapExactBaseForQuote(
+		poolId: string,
+		tokenObjectIn: string,
+		amountIn: bigint | number,
+		currentAddress: string,
+		clientOrderId: string | undefined = undefined,
+	): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		const [baseAsset, quoteAsset] = await this.getPoolTypeArgs(poolId);
+		// in this case, we assume that the tokenIn--tokenOut always exists.
+		const [base_coin_ret, quote_coin_ret, _amount] = txb.moveCall({
+			typeArguments: [baseAsset, quoteAsset],
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::swap_exact_base_for_quote`,
+			arguments: [
+				txb.object(poolId),
+				txb.pure(clientOrderId ?? this.#nextClientOrderId()),
+				txb.object(this.#checkAccountCap()),
+				txb.object(String(amountIn)),
+				txb.object(tokenObjectIn),
+				txb.moveCall({
+					typeArguments: [quoteAsset],
+					target: `0x2::coin::zero`,
+					arguments: [],
+				}),
+				txb.object(SUI_CLOCK_OBJECT_ID),
+			],
+		});
+		txb.transferObjects([base_coin_ret], txb.pure(currentAddress));
+		txb.transferObjects([quote_coin_ret], txb.pure(currentAddress));
+		return txb;
+	}
+
+	/**
+	 * @description: cancel an order
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param orderId orderId of a limit order, you can find them through function query.list_open_orders eg: "0"
+	 */
+	async cancelOrder(poolId: string, orderId: string): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::cancel_order`,
+			arguments: [txb.object(poolId), txb.pure(orderId), txb.object(this.#checkAccountCap())],
+		});
+		return txb;
+	}
+
+	/**
+	 * @description: Cancel all limit orders under a certain account capacity
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 */
+	async cancelAllOrders(poolId: string): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::cancel_all_orders`,
+			arguments: [txb.object(poolId), txb.object(this.#checkAccountCap())],
+		});
+		return txb;
+	}
+
+	/**
+	 * @description: batch cancel order
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param orderIds array of order ids you want to cancel, you can find your open orders by query.list_open_orders eg: ["0", "1", "2"]
+	 */
+	async batchCancelOrder(poolId: string, orderIds: string[]): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::batch_cancel_order`,
+			arguments: [txb.object(poolId), txb.pure(orderIds), txb.object(this.#checkAccountCap())],
+		});
+		return txb;
+	}
+
+	/**
+	 * @param poolId Object id of pool, created after invoking createPool, eg: "0xcaee8e1c046b58e55196105f1436a2337dcaa0c340a7a8c8baf65e4afb8823a4"
+	 * @param orderIds array of expired order ids to clean, eg: ["0", "1", "2"]
+	 * @param orderOwners array of Order owners, should be the owner addresses from the account capacities which placed the orders
+	 */
+	async cleanUpExpiredOrders(
+		poolId: string,
+		orderIds: string[],
+		orderOwners: string[],
+	): Promise<TransactionBlock> {
+		const txb = new TransactionBlock();
+		txb.moveCall({
+			typeArguments: await this.getPoolTypeArgs(poolId),
+			target: `${PACKAGE_ID}::${MODULE_CLOB}::clean_up_expired_orders`,
+			arguments: [
+				txb.object(poolId),
+				txb.object(SUI_CLOCK_OBJECT_ID),
+				txb.pure(orderIds),
+				txb.pure(orderOwners),
+			],
+		});
 		return txb;
 	}
 
@@ -163,6 +489,16 @@ export class DeepBookClient {
 			baseAsset,
 			quoteAsset,
 		};
+	}
+
+	async getPoolTypeArgs(poolId: string): Promise<string[]> {
+		if (!this.#poolTypeArgsCache.has(poolId)) {
+			const { baseAsset, quoteAsset } = await this.getPoolInfo(poolId);
+			const typeArgs = [baseAsset, quoteAsset];
+			this.#poolTypeArgsCache.set(poolId, typeArgs);
+		}
+
+		return this.#poolTypeArgsCache.get(poolId)!;
 	}
 
 	/**
@@ -316,7 +652,7 @@ export class DeepBookClient {
 		return normalizeSuiObjectId(cap);
 	}
 
-	#checkCurrentAddress(recipientAddress: string): string {
+	#checkAddress(recipientAddress: string): string {
 		if (recipientAddress === DUMMY_ADDRESS) {
 			throw new Error('Current address cannot be DUMMY_ADDRESS');
 		}
@@ -329,5 +665,11 @@ export class DeepBookClient {
 			options: { showType: true },
 		});
 		return Coin.getCoinTypeArg(resp);
+	}
+
+	#nextClientOrderId() {
+		const id = this.clientOrderId;
+		this.clientOrderId += 1;
+		return id;
 	}
 }
