+++ sdk/deepbook/src/index.ts
@@ -2,5 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './client';
-export * from './transactions';
 export * from './router';
